api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
scikit-image==0.21.0

# Color Rerank - Variant detection
scikit-learn>=1.0.0

# Local Feature Rerank - OpenCV ORB
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Image processing
Pillow>=9.0.0
numpy>=1.21.0